+++ src/imio/smartweb/core/contents/sections/gallery/content.py
@@ -11,14 +11,6 @@ from zope import schema
 class ISectionGallery(ISection):
     """Marker interface and Dexterity Python Schema for SectionGallery"""
 
-    model.fieldset("layout", fields=["image_scale"])
-    image_scale = schema.Choice(
-        title=_("Image scale for images"),
-        default="affiche",
-        vocabulary="imio.smartweb.vocabulary.Scales",
-        required=True,
-    )
-
 
 @implementer(ISectionGallery)
 class SectionGallery(Section):