+++ src/imio/smartweb/core/contents/sections/gallery/views.py
@@ -8,6 +8,6 @@ from imio.smartweb.core.contents.sections.views import SectionView
 class GalleryView(SectionView, PhotoGallery):
     """Gallery Section view"""
 
-    def get_scale_url(self, item, scale):
+    def get_scale_url(self, item):
         request = self.request
-        return get_scale_url(item, request, "image", scale)
+        return get_scale_url(item, request, "image")